requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
pandas==2.3.3
sentence-transformers==5.2.3
transformers==5.3.0
scikit-learn==1.8.0
numpy==2.4.2
jsonschema==4.25.1
faiss-cpu==1.13.2